# packages/wos-api/openapi.yaml
openapi: 3.1.0
info:
  title: Woobe Outcome Server HTTP API
  version: 0.1.0-dev
  description: |
    HTTP contract implemented through Wave 07. The API exposes the same
    Application services used by embedded consumers. Namespace and Outcome
    scope are explicit in every domain path. Issues and Blockers are independent
    aggregates; compound coordination commands remain explicit and atomic.
servers:
  - url: http://127.0.0.1:8080
...
paths:
...
  /api/v1/namespaces/{namespace_id}/outcomes/{outcome_id}/issues:
    get:
      tags: [Issues]
      operationId: listIssues
      parameters:
        - $ref: "#/components/parameters/NamespaceID"
        - $ref: "#/components/parameters/OutcomeID"
        - $ref: "#/components/parameters/CorrelationID"
      responses:
        "200":
          description: Outcome-local Issues ordered deterministically.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueCollection"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Issues]
      operationId: createIssue
      parameters:
        - $ref: "#/components/parameters/NamespaceID"
        - $ref: "#/components/parameters/OutcomeID"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CorrelationID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIssueRequest"
      responses:
        "201":
          description: Issue created or replayed idempotently.
          headers:
            ETag:
              schema: {type: string}
            Location:
              schema: {type: string}
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueMutationResult"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/v1/namespaces/{namespace_id}/outcomes/{outcome_id}/issues/actions/report-with-blocker:
    post:
      tags: [Issues, Blockers]
      operationId: reportIssueWithBlocker
      description: Atomically creates an Issue and one Blocker caused by that Issue.
      parameters:
        - $ref: "#/components/parameters/NamespaceID"
        - $ref: "#/components/parameters/OutcomeID"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CorrelationID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportIssueWithBlockerRequest"
      responses:
        "201":
          description: Both aggregates were confirmed under one outcome_revision.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportIssueWithBlockerMutationResult"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/v1/namespaces/{namespace_id}/outcomes/{outcome_id}/issues/{issue_id}:
    get:
      tags: [Issues]
      operationId: getIssue
      parameters:
        - $ref: "#/components/parameters/NamespaceID"
        - $ref: "#/components/parameters/OutcomeID"
        - $ref: "#/components/parameters/IssueID"
        - $ref: "#/components/parameters/CorrelationID"
      responses:
        "200":
          description: Persisted Issue representation.
          headers:
            ETag:
              schema: {type: string}
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueReadResult"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Issues]
      operationId: updateIssue
      parameters:
        - $ref: "#/components/parameters/NamespaceID"
        - $ref: "#/components/parameters/OutcomeID"
        - $ref: "#/components/parameters/IssueID"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIssueRequest"
      responses:
        "200":
          description: Issue details updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueMutationResult"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
...
  /api/v1/namespaces/{namespace_id}/outcomes/{outcome_id}/issues/{issue_id}/actions/mark-duplicate:
    post:
      tags: [Issues]
      operationId: markIssueDuplicate
      parameters:
        - $ref: "#/components/parameters/NamespaceID"
        - $ref: "#/components/parameters/OutcomeID"
        - $ref: "#/components/parameters/IssueID"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkIssueDuplicateRequest"
      responses:
        "200":
          description: Issue marked as duplicate after cycle validation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueMutationResult"
        "409":
          $ref: "#/components/responses/Conflict"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
  /api/v1/namespaces/{namespace_id}/outcomes/{outcome_id}/issues/{issue_id}/actions/resolve-with-blockers:
    post:
      tags: [Issues, Blockers]
      operationId: resolveIssueAndBlockers
      description: |
        Atomically resolves one Issue and an explicit bounded list of Blockers
        caused by it. Every listed Blocker requires its own expected_version and
        release_confirmed=true. Unlisted Blockers are untouched.
      parameters:
        - $ref: "#/components/parameters/NamespaceID"
        - $ref: "#/components/parameters/OutcomeID"
        - $ref: "#/components/parameters/IssueID"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveIssueAndBlockersRequest"
      responses:
        "200":
          description: Explicit multi-aggregate resolution confirmed atomically.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResolveIssueAndBlockersMutationResult"
        "409":
          $ref: "#/components/responses/Conflict"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
  /api/v1/namespaces/{namespace_id}/outcomes/{outcome_id}/blockers:
    get:
      tags: [Blockers]
      operationId: listBlockers
      parameters:
        - $ref: "#/components/parameters/NamespaceID"
        - $ref: "#/components/parameters/OutcomeID"
        - $ref: "#/components/parameters/CorrelationID"
      responses:
        "200":
          description: Outcome-local Blockers ordered deterministically.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockerCollection"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Blockers]
      operationId: createBlocker
      parameters:
        - $ref: "#/components/parameters/NamespaceID"
        - $ref: "#/components/parameters/OutcomeID"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CorrelationID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBlockerRequest"
      responses:
        "201":
          description: Blocker created or replayed idempotently.
          headers:
            ETag:
              schema: {type: string}
            Location:
              schema: {type: string}
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockerMutationResult"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/v1/namespaces/{namespace_id}/outcomes/{outcome_id}/blockers/{blocker_id}:
    get:
      tags: [Blockers]
      operationId: getBlocker
      parameters:
        - $ref: "#/components/parameters/NamespaceID"
        - $ref: "#/components/parameters/OutcomeID"
        - $ref: "#/components/parameters/BlockerID"
        - $ref: "#/components/parameters/CorrelationID"
      responses:
        "200":
          description: Persisted Blocker representation.
          headers:
            ETag:
              schema: {type: string}
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockerReadResult"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Blockers]
      operationId: updateBlocker
      description: Wave 07 allows editing the description of an active Blocker.
      parameters:
        - $ref: "#/components/parameters/NamespaceID"
        - $ref: "#/components/parameters/OutcomeID"
        - $ref: "#/components/parameters/BlockerID"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBlockerRequest"
      responses:
        "200":
          description: Blocker description updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockerMutationResult"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
...
components:
  parameters:
    NamespaceID:
      name: namespace_id
      in: path
      required: true
      schema: {$ref: "#/components/schemas/ID"}
    OutcomeID:
      name: outcome_id
      in: path
      required: true
      schema: {$ref: "#/components/schemas/ID"}
...
    IssueID:
      name: issue_id
      in: path
      required: true
      schema: {$ref: "#/components/schemas/ID"}
    BlockerID:
      name: blocker_id
      in: path
      required: true
      schema: {$ref: "#/components/schemas/ID"}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
    IfMatch:
      name: If-Match
      in: header
      required: false
      description: |
        Strong ETag of the persisted aggregate, for example
        "work_item:0199e100-0000-7000-8000-000000000001:v5".
        Either If-Match or expected_version in the body is required for
        versioned mutations. If both are present they must agree.
      schema: {type: string}
    CorrelationID:
      name: X-Correlation-ID
      in: header
      required: false
      schema: {type: string}
...
  responses:
...
    NotFound:
      description: Resource not found in the explicit Namespace/Outcome scope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Conflict:
      description: Lifecycle, lease, criterion or idempotency conflict.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    PreconditionFailed:
      description: Persisted version did not satisfy the supplied precondition.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    PreconditionRequired:
      description: A required If-Match or expected_version precondition is absent.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    UnprocessableEntity:
      description: Request format is valid but a domain invariant is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    ID:
      type: string
      format: uuid
      pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
    Version:
      type: integer
      minimum: 1
...
    IssueSeverity:
      type: string
      enum: [critical, major, minor, informational]
...
    BlockerPropagation:
      type: string
      enum: [direct, subtree]
...
    CreateIssueRequest:
      type: object
      additionalProperties: false
      required: [title, severity]
      properties:
        title: {type: string}
        description: {type: string}
        severity: {$ref: "#/components/schemas/IssueSeverity"}
        affected_refs:
          type: array
          items: {$ref: "#/components/schemas/EntityRefRequest"}
    UpdateIssueRequest:
      type: object
      additionalProperties: false
      properties:
        expected_version: {$ref: "#/components/schemas/Version"}
        title: {type: string}
        description: {type: string}
        severity: {$ref: "#/components/schemas/IssueSeverity"}
        affected_refs:
          type: array
          items: {$ref: "#/components/schemas/EntityRefRequest"}
...
    CreateBlockerRequest:
      type: object
      additionalProperties: false
      required: [blocked_ref]
      properties:
        blocked_ref: {$ref: "#/components/schemas/EntityRefRequest"}
        cause_ref: {$ref: "#/components/schemas/EntityRefRequest"}
        external_cause: {$ref: "#/components/schemas/ExternalCause"}
        description: {type: string}
        propagation: {$ref: "#/components/schemas/BlockerPropagation"}
    UpdateBlockerRequest:
      type: object
      additionalProperties: false
      required: [description]
      properties:
        expected_version: {$ref: "#/components/schemas/Version"}
        description: {type: string}
...
    MarkIssueDuplicateRequest:
      type: object
      additionalProperties: false
      required: [duplicate_of_issue_id]
      properties:
        expected_version: {$ref: "#/components/schemas/Version"}
        duplicate_of_issue_id: {$ref: "#/components/schemas/ID"}
    ReportIssueWithBlockerRequest:
      type: object
      additionalProperties: false
      required: [issue, blocker]
      properties:
        issue:
          $ref: "#/components/schemas/CreateIssueRequest"
        blocker:
          type: object
          additionalProperties: false
          required: [blocked_ref]
          properties:
            blocked_ref: {$ref: "#/components/schemas/EntityRefRequest"}
            description: {type: string}
            propagation: {$ref: "#/components/schemas/BlockerPropagation"}
...
    ResolveIssueAndBlockersRequest:
      type: object
      additionalProperties: false
      required: [expected_issue_version, issue_resolution_summary, blockers]
      properties:
        expected_issue_version: {$ref: "#/components/schemas/Version"}
        issue_resolution_summary: {type: string}
        blockers:
          type: array
          minItems: 1
          maxItems: 64
          items: {$ref: "#/components/schemas/BlockerResolutionRequest"}
...
    Issue:
      type: object
      required: [id, scope, version, title, severity, affected_refs, lifecycle, reported_by, created_at, updated_at]
      properties:
        id: {$ref: "#/components/schemas/ID"}
        scope: {type: object}
        version: {$ref: "#/components/schemas/Version"}
        title: {type: string}
        description: {type: string}
        severity: {$ref: "#/components/schemas/IssueSeverity"}
        affected_refs:
          type: array
          items: {type: object}
        lifecycle: {$ref: "#/components/schemas/IssueLifecycle"}
        reported_by: {$ref: "#/components/schemas/ActorRef"}
        resolution_summary: {type: string}
        duplicate_of_issue_id: {$ref: "#/components/schemas/ID"}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
...
    IssueMutationResult:
      allOf:
        - $ref: "#/components/schemas/MutationMetadata"
        - type: object
          required: [value]
          properties:
            value: {$ref: "#/components/schemas/Issue"}
    BlockerMutationResult:
      allOf:
        - $ref: "#/components/schemas/MutationMetadata"
        - type: object
          required: [value]
          properties:
            value: {$ref: "#/components/schemas/Blocker"}
    ReportIssueWithBlockerMutationResult:
      allOf:
        - $ref: "#/components/schemas/MutationMetadata"
        - type: object
          required: [value]
          properties:
            value:
              type: object
              required: [issue, blocker]
              properties:
                issue: {$ref: "#/components/schemas/Issue"}
                blocker: {$ref: "#/components/schemas/Blocker"}
    ResolveIssueAndBlockersMutationResult:
      allOf:
        - $ref: "#/components/schemas/MutationMetadata"
        - type: object
          required: [value]
          properties:
            value:
              type: object
              required: [issue, blockers]
              properties:
                issue: {$ref: "#/components/schemas/Issue"}
                blockers:
                  type: array
                  items: {$ref: "#/components/schemas/Blocker"}
...
    IssueReadResult:
      type: object
      required: [value, outcome_revision]
      properties:
        value: {$ref: "#/components/schemas/Issue"}
        outcome_revision: {type: integer, minimum: 1}
    BlockerReadResult:
      type: object
      required: [value, outcome_revision]
      properties:
        value: {$ref: "#/components/schemas/Blocker"}
        outcome_revision: {type: integer, minimum: 1}
...
    IssueCollection:
      type: object
      required: [items, outcome_revision]
      properties:
        items:
          type: array
          items: {$ref: "#/components/schemas/Issue"}
        outcome_revision: {type: integer, minimum: 1}
    BlockerCollection:
      type: object
      required: [items, outcome_revision]
      properties:
        items:
          type: array
          items: {$ref: "#/components/schemas/Blocker"}
        outcome_revision: {type: integer, minimum: 1}
...